ft_printf full version

+++ ft_puthex.c
@@ -6,13 +6,13 @@
 /*   By: oumondad <oumondad@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/01 12:28:17 by oumondad          #+#    #+#             */
-/*   Updated: 2023/12/01 16:11:21 by oumondad         ###   ########.fr       */
+/*   Updated: 2023/12/03 19:01:22 by oumondad         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-static int	hexlen(unsigned long nb)
+static int	hexlen(unsigned int nb)
 {
 	int	i;
 
@@ -25,9 +25,9 @@ static int	hexlen(unsigned long nb)
 	return (i);
 }
 
-int	ft_puthex(unsigned long nb, char *base)
+int	ft_puthex(unsigned int nb, char *base)
 {
-	if(nb == 0)
+	if (nb == 0)
 	{
 		ft_putchar('0');
 		return (1);
